tests nearly pass with python helper functions refactored
